+++ flow/wrapper.flow.js
@@ -28,6 +28,7 @@ declare interface BaseWrapper {
   name(): string | void;
   props(key?: string): { [name: string]: any } | any | void;
   text(): string | void;
+  selector: Selector | void;
   setData(data: Object): void;
   setMethods(methods: Object): void;
   setValue(value: any): void;
@@ -40,6 +41,5 @@ declare interface BaseWrapper {
 
 declare type WrapperOptions = {
   // eslint-disable-line no-undef
-  attachedToDocument?: boolean,
-  sync?: boolean
+  attachedToDocument?: boolean
 }