+++ packages/server-test-utils/src/index.js
@@ -2,8 +2,4 @@ import renderToString from './renderToString'
 import render from './render'
 import config from './config'
 
-export default {
-  renderToString,
-  config,
-  render
-}
+export { renderToString, config, render }