+++ test/specs/wrapper-array/name.spec.js
@@ -1,6 +1,6 @@
 import { describeWithShallowAndMount } from '~resources/utils'
 import { compileToFunctions } from 'vue-template-compiler'
-import '~vue/test-utils'
+import '@vue/test-utils'
 
 describeWithShallowAndMount('name', mountingMethod => {
   it('throws an error when called on a WrapperArray', () => {